Add a voltage to DAC value mapping macro.
Voltages in the DACs are represented by a 16-bit unsigned integer with
0x0000 being the minimum absolute voltage (presently -10 V) and 0xFFFF
the absolute maximum (presently +10 V). This fits well in along the line
VOLTAGE = 3.05179e-4 * CODE - 10.0 with 1 mV precision. A macro has been
created to do this transformation. The macro can be used for different
slopes and intercepts by using its expanded version.