Change `compile` to return a ref to the instruction itself
The `compile` function of `Instruction` previously returned a &[u32] view
of the instruction. However this capability was already served by the
`view` function of the same trait. This commit changes `compile` to
return the instruction itself so that it can be chained with the other
functions (such as `to_bytevec` which will ultimately be used by the
instruction processor).