Split long delays into a slow and a fast delay
If a large delay is requested ArC2 will now split the delay into two
different delays. A large one that holds the bulk of the delaying and
a smaller zero-length delay (which is really 320 ns). That way if the
delay instruction is the last in the instruction buffer then it will
be terminated on a small delay. This is necessary because if a large
delay is the last command in the buffer the FIFO will erroneously
look empty while ArC2 is busy delaying. Adding a very fast delay at the
end of the instruction buffer will prevent that from happening. The
threshold is set fairly low for now but will be adjusted as we go.