Add a function for singular crosspoint pulsing
This commit exposes singular pulsing functionality on the top level
interface. The pulsing functionality is split into a fast and a slow
path. In most use cases the high speed function will be used unless
pulses longer than 500 ms are needed. In this case the high speed drivers
of ArC2 will be used in a differential biasing mode.