`SetDAC::from_channels` should not overwrite previously set clusters
The initial approach when building SetDAC::from_channels, which was to
group channels per cluster, was quite problematic as it is not possible
to set a single channel within a cluster. Therefore, even though the
instructions that were emitted are correct they would overwrite the
setting of the unselected channels on a per-instruction basis.
This commit alters the original implementation of from_channels to only
group clusters when _all four_ offsets are identical.