Add an option to preset high speed state before pulsing
The fast pulsing functions now have an additional argument that allows
for presetting the state of the HS drivers before the actual pulsing
occurs. This is implemented with a zero width HS Pulse that's applied
before the DACs are configured and essentially prepares the state of the
high speed drivers before the actual pulsing. This ensures that channels
are not improperly biased between successive pulses when setting up
the DACs and/or transitioning from VoltArb to HiSpeed and vice versa.