Implement chunked FPGA writes
The FPGA FIFO has a finite amount of space for instructions, which means
that user-facing code must keep track of the size of the FIFO when
schedulling instructions for execution. This commit changes how the
computer-side instruction buffer operates by splitting it into
FPGA-compatible chunks when `execute` is called. There is some penalty
when the computer-side buffer has more instructions than the FPGA FIFO
can handle as an implicit `wait` is triggered to make sure that the FIFO
is fully consumed before new instructions are loaded.