Add non-locking variants of `busy()` and `wait()` for internal use
When processing the instruction buffer, the inner loop of `execute()`
needs to wait for the FPGA FIFO to be consumed otherwise instructions
will be dropped. This was typically done with the main `busy()` and
`wait()` function. As these are typically meant to be user-facing code
they require a lock to `bl::Device` to proceed. For the case of execute
this is not necessary as it already holds a lock to the device for
writing. Trying to relock `bl::Device` from within `wait()` will end in
a deadlock. For this reason non-locking variants of `busy()` and `wait()`
have been created purely for internal use. User-facing code should never
have to access these as it will never have to read/write raw buffers from
the FPGA.