Add special handling for ramping when `num_pulses` is 0
The basic ramp behaviour is typically a pulse followed by an optional
read. Sometimes it is useful to only have reads instead of pulses. Since
`num_pulses` indicates how many programming pulses are issued setting it
to 0 will forgo the application of programming pulses and instead skip
directly to the read-out phase. This needs `ReadAfter::Pulse`.