Centralise memory access facilities outside MemMan
Instead of functions having to read and clear memory directly new functions
are provided to wrap this instead. Functions `Instrument::make_chunk`,
and `Instrument::read_chunk` can now be used to create and read an
allocated chunk. This way it is guaranteed that chunks will be freed once
they are picked from the memory.