Initial support for flag addresses
This commits adds support for addresses that hold "operation completed"
flags. These will be used by the firmware to notify the donwstream user
of the library that a result has been committed to memory. It is the
responsibility of the user to clear the flag after the value has been
retrieved. This functionality is gated behind a cargo feature because
more testing is currently required.