Make `pick_one` use the flag address functionality if available
This is further work for the "flag addresses" functionality to ensure
that values are populated before being retrieved. This is necessary for
longer, more complicated sequences as it allows to have feedback from the
process before continuing. It also allows the interface to pick up values
as they are becoming available instead of waiting for the EFM to return
to a non-busy status.

It's still experimental so it's gated behind the `flag_addresses` feature.