Expose a function to control the ArC2 digital I/Os
Expose digital I/O control to libarc2. Previously the only interaction
the tool had with the "Update Logic" command was when initialising the
tool or when switching between header and internal control. This commit
adds function `set_logic` that turns the selected digital outputs high
or low arbitrarily. To make things simpler the `IOMask` required can now
also be created through a `u32` instead of messing with the register
internals directly (it is, after all, a bit vector backed by `u32`s).
This also makes the python wrapper simpler.