Remove ndarray dependency
The only use of ndarray was in translating internal buffers to Python
arrays for consumption by pyarc2. This has now been moved to pyarc2
itself which is the more natural for the conversion to take place.
Apart from that the only other use of ndarray was for constructing
the equivalent of numpy.range(start, stop, step) when generating
ramps. This has been replaced with equivalent code for step generation
which is essentially identical to the way ndarray does it.

It is possible that ndarray might get reintroduced at some point in
future as the library develops (so it might actually be sooner rather
than later). In that case some of the code might get refactored to
use it efficiently. Regardless of that the buffer-to-ndarray
conversions will not return.