Drop comparisons with base voltage in `from_channels` if clear is false
SetDAC::from_channels had an unexpected siee-effect when clear is set
to false. The SetDAC::from_channels compile would check channels in
sets of four (one half-channel) against the base voltage of the function
to determine if a duplicate instruction would be emitted. So if the base
voltage is 0x80008000 then if all four consecutive channels of a cluster
are selected and the chosen voltage is 0x80008000 the corresponding LD
VOLT instruction would be dropped because the channels are there already.
This check happened even when clear was false however this has the
unintended side-effect that if 0.0 V (or the base voltage) is passed as
an argument for four consecutive clusters this would lead to no instruction
being emitted. This commit ensures that no checks of that kind are done
when clear is false. Follow the explicit instructions defined in input
even though this might lead to some duplicate instructions in very
specific cirumstances.